# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ANVISA-RE Monitor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/resolucoes:
    get:
      operationId: listResolucoes
      tags: [resolucoes]
      summary: Listar Resoluções Específicas com filtros e paginação
      parameters:
        - name: q
          in: query
          schema:
            type: ["string", "null"]
          description: Busca full-text (nome do produto, princípio ativo, fabricante, ementa)
        - name: status
          in: query
          schema:
            type: ["string", "null"]
          description: "Filtro por status: vigente, revogada, encerrada, em_analise"
        - name: tipo_produto
          in: query
          schema:
            type: ["string", "null"]
          description: Filtro por categoria de produto
        - name: tipo_acao
          in: query
          schema:
            type: ["string", "null"]
          description: Filtro por tipo de ação sanitária
        - name: data_inicio
          in: query
          schema:
            type: ["string", "null"]
          description: "Data de publicação início (ISO 8601: YYYY-MM-DD)"
        - name: data_fim
          in: query
          schema:
            type: ["string", "null"]
          description: "Data de publicação fim (ISO 8601: YYYY-MM-DD)"
        - name: numero_re
          in: query
          schema:
            type: ["string", "null"]
          description: Número exato da RE
        - name: numero_registro_anvisa
          in: query
          schema:
            type: ["string", "null"]
          description: Número de registro ANVISA
        - name: fabricante_cnpj
          in: query
          schema:
            type: ["string", "null"]
          description: CNPJ do fabricante
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
          description: Página (default 1)
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
          description: Itens por página (default 20, max 100)
        - name: order_by
          in: query
          schema:
            type: ["string", "null"]
          description: "Campo de ordenação (default: data_publicacao)"
        - name: order_dir
          in: query
          schema:
            type: ["string", "null"]
          description: "Direção: asc ou desc"
      responses:
        "200":
          description: Lista paginada de REs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResolucoes"

    post:
      operationId: createResolucao
      tags: [resolucoes]
      summary: Criar nova Resolução Específica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResolucaoBody"
      responses:
        "201":
          description: RE criada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resolucao"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/resolucoes/exportar:
    get:
      operationId: exportarResolucoes
      tags: [resolucoes]
      summary: Exportar resultados de busca em CSV, XLSX ou PDF
      parameters:
        - name: formato
          in: query
          required: true
          schema:
            type: string
          description: "Formato de exportação: csv, xlsx, pdf"
        - name: q
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: tipo_produto
          in: query
          schema:
            type: ["string", "null"]
        - name: tipo_acao
          in: query
          schema:
            type: ["string", "null"]
        - name: data_inicio
          in: query
          schema:
            type: ["string", "null"]
        - name: data_fim
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Arquivo exportado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
        "400":
          description: Formato inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/resolucoes/{id}:
    get:
      operationId: getResolucao
      tags: [resolucoes]
      summary: Detalhe completo de uma RE
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalhe da RE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resolucao"
        "404":
          description: RE não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateResolucao
      tags: [resolucoes]
      summary: Atualizar RE existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResolucaoBody"
      responses:
        "200":
          description: RE atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resolucao"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: RE não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteResolucao
      tags: [resolucoes]
      summary: Exclusão lógica (soft delete) de uma RE
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RE removida com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: RE não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/resolucoes/{id}/historico:
    get:
      operationId: getResolucaoHistorico
      tags: [resolucoes]
      summary: Histórico de alterações de uma RE
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de versões/alterações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoricoEntry"
        "404":
          description: RE não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Resolucao:
      type: object
      properties:
        id:
          type: string
        numero_re:
          type: string
        data_publicacao:
          type: string
        data_vigencia_inicio:
          type: string
        data_vigencia_fim:
          type: ["string", "null"]
        status:
          type: string
          description: "vigente | revogada | encerrada | em_analise"
        tipo_produto:
          type: string
        tipo_acao:
          type: array
          items:
            type: string
        nome_produto:
          type: string
        principio_ativo:
          type: ["string", "null"]
        fabricante_nome:
          type: ["string", "null"]
        fabricante_cnpj:
          type: ["string", "null"]
        numero_registro_anvisa:
          type: ["string", "null"]
        ementa:
          type: string
        link_documento_oficial:
          type: ["string", "null"]
        arquivo_pdf_path:
          type: ["string", "null"]
        criado_por:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
        deleted_at:
          type: ["string", "null"]
      required:
        - id
        - numero_re
        - data_publicacao
        - data_vigencia_inicio
        - status
        - tipo_produto
        - tipo_acao
        - nome_produto
        - ementa
        - created_at
        - updated_at
...
    PaginatedResolucoes:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Resolucao"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - total_pages
    CreateResolucaoBody:
      type: object
      properties:
        numero_re:
          type: string
        data_publicacao:
          type: string
        data_vigencia_inicio:
          type: string
        data_vigencia_fim:
          type: ["string", "null"]
        status:
          type: string
        tipo_produto:
          type: string
        tipo_acao:
          type: array
          items:
            type: string
        nome_produto:
          type: string
        principio_ativo:
          type: ["string", "null"]
        fabricante_nome:
          type: ["string", "null"]
        fabricante_cnpj:
          type: ["string", "null"]
        numero_registro_anvisa:
          type: ["string", "null"]
        ementa:
          type: string
        link_documento_oficial:
          type: ["string", "null"]
        arquivo_pdf_path:
          type: ["string", "null"]
      required:
        - numero_re
        - data_publicacao
        - data_vigencia_inicio
        - status
        - tipo_produto
        - tipo_acao
        - nome_produto
        - ementa
    UpdateResolucaoBody:
      type: object
      properties:
        data_vigencia_fim:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        tipo_produto:
          type: ["string", "null"]
        tipo_acao:
          type: ["array", "null"]
          items:
            type: string
        nome_produto:
          type: ["string", "null"]
        principio_ativo:
          type: ["string", "null"]
        fabricante_nome:
          type: ["string", "null"]
        fabricante_cnpj:
          type: ["string", "null"]
        numero_registro_anvisa:
          type: ["string", "null"]
        ementa:
          type: ["string", "null"]
        link_documento_oficial:
          type: ["string", "null"]
        arquivo_pdf_path:
          type: ["string", "null"]
        justificativa_alteracao:
          type: ["string", "null"]
    HistoricoEntry:
      type: object
      properties:
        id:
          type: string
        resolucao_id:
          type: string
        campo_alterado:
          type: ["string", "null"]
        valor_anterior:
          type: ["string", "null"]
        valor_novo:
          type: ["string", "null"]
        justificativa:
          type: ["string", "null"]
        alterado_por:
          type: ["string", "null"]
        alterado_por_nome:
          type: ["string", "null"]
        fonte:
          type: string
        created_at:
          type: string
        dados_anteriores:
          type: ["object", "null"]
        dados_novos:
          type: ["object", "null"]
      required:
        - id
        - resolucao_id
        - fonte
        - created_at
...
    ExportResponse:
      type: object
      properties:
        url:
          type: string
        formato:
          type: string
        total_registros:
          type: integer
        filename:
          type: string
      required:
        - url
        - formato
        - total_registros
        - filename
...